Refactor llm_utils
