+++ solib/Experiment.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 from solib.data.loading import Dataset
 from solib.utils import str_config, write_json, dump_config, random
-from solib.llm_utils import parallelized_call
+from solib.utils import parallelized_call
 from solib.protocols.protocols import (
     Blind,
     Propaganda,
