+++ solib/protocols/judges/JustAskProbabilityJudge.py
@@ -1,5 +1,5 @@
 import logging
-from solib.llm_utils import parallelized_call
+from solib.utils import parallelized_call
 from solib.datatypes import Question, Answer, Prob
 from solib.tool_use.tool_rendering import TRUST_TOOL_USE_PROMPT
 from solib.protocols.abstract import Judge
