+++ solib/tool_use/tool_use.py
@@ -1,15 +1,14 @@
-from typing import Callable, Dict, List, Literal
+from typing import Callable, Dict, List
 import torch
 
 torch.set_grad_enabled(False)
-from transformers import AutoTokenizer, AutoModelForCausalLM, StoppingCriteria, StoppingCriteriaList
+from transformers import AutoTokenizer, StoppingCriteria
 import re
 import json
 
-from langchain_core.messages import ToolMessage, SystemMessage, BaseMessage
+from langchain_core.messages import ToolMessage, BaseMessage
 
 from solib.tool_use.tool_rendering import get_structured_tools, render_tool_call_result, TOOL_CALL_START_TAG, TOOL_CALL_END_TAG
-from solib.utils import apply, apply_async
